social_distribution/requirements.txt
asgiref==3.5.0
Django==4.0.1
django-nested-admin==3.4.0
djangorestframework==3.13.1
Pillow==9.0.1
python-monkey-business==1.0.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
